+++ styles/src/base_layers.ts
@@ -1,4 +1,9 @@
-import { LayerSpecification } from "@maplibre/maplibre-gl-style-spec";
+import {
+  DataDrivenPropertyValueSpecification,
+  ExpressionSpecification,
+  LayerSpecification,
+} from "@maplibre/maplibre-gl-style-spec";
+import { get_country_name, get_multiline_name } from "./language";
 import { Theme } from "./themes";
 
 export function nolabels_layers(
@@ -1634,7 +1639,12 @@ export function nolabels_layers(
   ];
 }
 
-export function labels_layers(source: string, t: Theme): LayerSpecification[] {
+export function labels_layers(
+  source: string,
+  t: Theme,
+  lang: string,
+  script?: string,
+): LayerSpecification[] {
   return [
     {
       id: "physical_line_waterway_label",
@@ -1646,7 +1656,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       layout: {
         "symbol-placement": "line",
         "text-font": ["Noto Sans Regular"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": 12,
         "text-letter-spacing": 0.3,
       },
@@ -1662,7 +1675,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       filter: ["any", ["==", "pmap:kind", "peak"]],
       layout: {
         "text-font": ["Noto Sans Italic"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": ["interpolate", ["linear"], ["zoom"], 10, 8, 16, 12],
         "text-letter-spacing": 0.1,
         "text-max-width": 9,
@@ -1683,7 +1699,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
         "symbol-sort-key": ["get", "pmap:min_zoom"],
         "symbol-placement": "line",
         "text-font": ["Noto Sans Regular"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": 12,
       },
       paint: {
@@ -1713,7 +1732,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       ],
       layout: {
         "text-font": ["Noto Sans Medium"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": ["interpolate", ["linear"], ["zoom"], 3, 10, 10, 12],
         "text-letter-spacing": 0.1,
         "text-max-width": 9,
@@ -1731,7 +1753,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       filter: ["any", ["in", "pmap:kind", "lake", "water"]],
       layout: {
         "text-font": ["Noto Sans Medium"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": ["interpolate", ["linear"], ["zoom"], 3, 0, 6, 12, 10, 12],
         "text-letter-spacing": 0.1,
         "text-max-width": 9,
@@ -1754,7 +1779,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
         "symbol-sort-key": ["get", "pmap:min_zoom"],
         "symbol-placement": "line",
         "text-font": ["Noto Sans Regular"],
-        "text-field": ["get", "name"],
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-size": 12,
       },
       paint: {
@@ -1771,7 +1799,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       filter: ["==", "pmap:kind", "neighbourhood"],
       layout: {
         "symbol-sort-key": ["get", "pmap:min_zoom"],
-        "text-field": "{name}",
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-font": ["Noto Sans Regular"],
         "text-max-width": 7,
         "text-letter-spacing": 0.1,
@@ -1816,7 +1847,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       layout: {
         "icon-image": ["step", ["zoom"], "townspot", 8, ""],
         "icon-size": 0.7,
-        "text-field": "{name}",
+        "text-field": get_multiline_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-font": [
           "case",
           ["<=", ["get", "pmap:min_zoom"], 5],
@@ -1930,7 +1964,7 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
           ["zoom"],
           ["get", "name:short"],
           6,
-          ["get", "name"],
+          get_multiline_name(lang, script) as ExpressionSpecification,
         ],
         "text-font": ["Noto Sans Regular"],
         "text-size": ["interpolate", ["linear"], ["zoom"], 3, 11, 7, 16],
@@ -1952,7 +1986,10 @@ export function labels_layers(source: string, t: Theme): LayerSpecification[] {
       filter: ["==", "pmap:kind", "country"],
       layout: {
         "symbol-sort-key": ["get", "pmap:min_zoom"],
-        "text-field": "{name}",
+        "text-field": get_country_name(
+          lang,
+          script,
+        ) as DataDrivenPropertyValueSpecification<string>,
         "text-font": ["Noto Sans Medium"],
         "text-size": [
           "interpolate",