+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Styles v4.0.0-alpha.0
+------
+- Add lang and script parameters to TypeScript style generation [#275]
+
 Tiles v4.0.0-alpha.3
 ------
 - Replace Natural Earth places at low zooms with OSM [#289]